refactor: rename TryRequestButton to TestRequestButton
