+++ packages/api-reference/src/components/ApiClientModal.vue
@@ -1,21 +1,18 @@
 <script setup lang="ts">
-import type { Spec } from '@scalar/oas-utils'
 import { onMounted, ref } from 'vue'
 
 import { apiClientBus } from './api-client-bus'
 
 const props = defineProps<{
-  parsedSpec: Spec
   proxyUrl?: string
 }>()
 
 const el = ref<HTMLDivElement | null>(null)
 
 onMounted(async () => {
-  if (!el.value) return
-
-  console.log(props.parsedSpec)
-  console.log(props.proxyUrl)
+  if (!el.value) {
+    return
+  }
 
   const { createScalarApiClient } = await import('@scalar/api-client-modal')
   const { open } = await createScalarApiClient(el.value, {
